move router to /nlu
