+++ src/bp/core/routers/bots/content.ts
@@ -1,35 +1,18 @@
 import { ContentElement, Logger } from 'botpress/sdk'
-import * as sdk from 'botpress/sdk'
 import { LibraryElement } from 'common/typings'
-import { EntityDefCreateSchema, IntentDefCreateSchema } from 'common/validation'
 import { GhostService } from 'core/services'
 import AuthService, { TOKEN_AUDIENCE } from 'core/services/auth/auth-service'
-import { InvalidOperationError } from 'core/services/auth/errors'
 import { CMSService } from 'core/services/cms'
-import { CmsImportSchema, DefaultSearchParams } from 'core/services/cms'
-import { EntityService } from 'core/services/nlu/entities-service'
-import { IntentService } from 'core/services/nlu/intent-service'
+import { DefaultSearchParams } from 'core/services/cms'
 import { WorkspaceService } from 'core/services/workspace-service'
 import { RequestHandler, Router } from 'express'
-import { validate } from 'joi'
 import _ from 'lodash'
-import moment from 'moment'
-import multer from 'multer'
-import yn from 'yn'
 
 import { CustomRouter } from '../customRouter'
 import { checkTokenHeader, needPermissions } from '../util'
 
 const CONTENT_FOLDER = 'content-elements'
 const LIBRARY_FILE = 'library.json'
-
-const removeSlotsFromUtterances = (utterances: { [key: string]: any }, slotNames: string[]) =>
-  _.fromPairs(
-    Object.entries(utterances).map(([key, val]) => {
-      const regex = new RegExp(`\\[([^\\[\\]\\(\\)]+?)\\]\\((${slotNames.join('|')})\\)`, 'gi')
-      return [key, val.map(u => u.replace(regex, '$1'))]
-    })
-  )
 
 export class ContentRouter extends CustomRouter {
   private _checkTokenHeader: RequestHandler
@@ -40,9 +23,7 @@ export class ContentRouter extends CustomRouter {
     private authService: AuthService,
     private cms: CMSService,
     private workspaceService: WorkspaceService,
-    private ghost: GhostService,
-    private entityService: EntityService,
-    private intentService: IntentService
+    private ghost: GhostService
   ) {
     super('Content', logger, Router({ mergeParams: true }))
     this._needPermissions = needPermissions(this.workspaceService)
@@ -241,212 +222,6 @@ export class ContentRouter extends CustomRouter {
         await ghost.upsertFile(CONTENT_FOLDER, LIBRARY_FILE, JSON.stringify([...contentIds, contentId], undefined, 2))
 
         res.sendStatus(200)
-      })
-    )
-
-    this.router.get(
-      '/intents',
-      this._checkTokenHeader,
-      this._needPermissions('write', 'bot.content'),
-      this.asyncMiddleware(async (req, res) => {
-        const { botId } = req.params
-        const intentDefs = await this.intentService.getIntents(botId)
-        res.send(intentDefs)
-      })
-    )
-
-    this.router.get(
-      '/intents/:intent',
-      this._checkTokenHeader,
-      this._needPermissions('write', 'bot.content'),
-      this.asyncMiddleware(async (req, res) => {
-        const { botId, intent } = req.params
-        const intentDef = await this.intentService.getIntent(botId, intent)
-        res.send(intentDef)
-      })
-    )
-
-    this.router.post(
-      '/intents/:intent/delete',
-      this._checkTokenHeader,
-      this._needPermissions('write', 'bot.content'),
-      this.asyncMiddleware(async (req, res) => {
-        const { botId, intent } = req.params
-        try {
-          await this.intentService.deleteIntent(botId, intent)
-          res.sendStatus(204)
-        } catch (err) {
-          this.logger
-            .forBot(botId)
-            .attachError(err)
-            .error('Could not delete intent')
-          res.status(400).send(err.message)
-        }
-      })
-    )
-
-    this.router.post(
-      '/intents',
-      this._checkTokenHeader,
-      this._needPermissions('write', 'bot.content'),
-      this.asyncMiddleware(async (req, res) => {
-        const { botId } = req.params
-        try {
-          const intentDef = await validate(req.body, IntentDefCreateSchema, {
-            stripUnknown: true
-          })
-
-          await this.intentService.saveIntent(botId, intentDef)
-
-          res.sendStatus(200)
-        } catch (err) {
-          this.logger
-            .forBot(botId)
-            .attachError(err)
-            .warn('Cannot create intent')
-          res.status(400).send(err.message)
-        }
-      })
-    )
-
-    this.router.post(
-      '/intents/:intentName',
-      this._checkTokenHeader,
-      this._needPermissions('write', 'bot.content'),
-      this.asyncMiddleware(async (req, res) => {
-        const { botId, intentName } = req.params
-        try {
-          await this.intentService.updateIntent(botId, intentName, req.body)
-          res.sendStatus(200)
-        } catch (err) {
-          this.logger
-            .forBot(botId)
-            .attachError(err)
-            .error('Could not update intent')
-          res.sendStatus(400)
-        }
-      })
-    )
-
-    this.router.get(
-      '/entities',
-      this._checkTokenHeader,
-      this._needPermissions('write', 'bot.content'),
-      this.asyncMiddleware(async (req, res) => {
-        const { botId } = req.params
-        const { ignoreSystem } = req.query
-
-        const entities = await this.entityService.getEntities(botId)
-        const mapped = entities.map(x => ({ ...x, label: `${x.type}.${x.name}` }))
-
-        res.json(yn(ignoreSystem) ? mapped.filter(x => x.type !== 'system') : mapped)
-      })
-    )
-
-    this.router.get(
-      '/entities/:entityName',
-      this._checkTokenHeader,
-      this._needPermissions('write', 'bot.content'),
-      this.asyncMiddleware(async (req, res) => {
-        const { botId, entityName } = req.params
-        try {
-          const entity = await this.entityService.getEntity(botId, entityName)
-          res.send(entity)
-        } catch (err) {
-          this.logger
-            .forBot(botId)
-            .attachError(err)
-            .error(`Could not get entity ${entityName}`)
-          res.send(400)
-        }
-      })
-    )
-
-    this.router.post(
-      '/entities',
-      this._checkTokenHeader,
-      this._needPermissions('write', 'bot.content'),
-      this.asyncMiddleware(async (req, res) => {
-        const { botId } = req.params
-        try {
-          const entityDef = (await validate(req.body, EntityDefCreateSchema, {
-            stripUnknown: true
-          })) as sdk.NLU.EntityDefinition
-
-          await this.entityService.saveEntity(botId, entityDef)
-
-          res.sendStatus(200)
-        } catch (err) {
-          this.logger
-            .forBot(botId)
-            .attachError(err)
-            .warn('Cannot create entity')
-          res.status(400).send(err.message)
-        }
-      })
-    )
-
-    this.router.post(
-      '/entities/:id',
-      this._checkTokenHeader,
-      this._needPermissions('write', 'bot.content'),
-      this.asyncMiddleware(async (req, res) => {
-        const { botId, id } = req.params
-        try {
-          const entityDef = (await validate(req.body, EntityDefCreateSchema, {
-            stripUnknown: true
-          })) as sdk.NLU.EntityDefinition
-
-          await this.entityService.updateEntity(botId, id, entityDef)
-          res.sendStatus(200)
-        } catch (err) {
-          this.logger
-            .forBot(botId)
-            .attachError(err)
-            .error('Could not update entity')
-          res.status(400).send(err.message)
-        }
-      })
-    )
-
-    this.router.post(
-      '/entities/:id/delete',
-      this._checkTokenHeader,
-      this._needPermissions('write', 'bot.content'),
-      this.asyncMiddleware(async (req, res) => {
-        const { botId, id } = req.params
-        try {
-          await this.entityService.deleteEntity(botId, id)
-
-          const affectedIntents = (await this.intentService.getIntents(botId)).filter(intent =>
-            intent.slots.some(slot => slot.entities.includes(id))
-          )
-
-          await Promise.map(affectedIntents, intent => {
-            const [affectedSlots, unaffectedSlots] = _.partition(intent.slots, slot => slot.entities.includes(id))
-            const [slotsToDelete, slotsToKeep] = _.partition(affectedSlots, slot => slot.entities.length === 1)
-            const updatedIntent = {
-              ...intent,
-              slots: [
-                ...unaffectedSlots,
-                ...slotsToKeep.map(slot => ({ ...slot, entities: _.without(slot.entities, id) }))
-              ],
-              utterances: removeSlotsFromUtterances(
-                intent.utterances,
-                slotsToDelete.map(slot => slot.name)
-              )
-            }
-            return this.intentService.saveIntent(botId, updatedIntent)
-          })
-
-          res.sendStatus(204)
-        } catch (err) {
-          this.logger
-            .forBot(botId)
-            .attachError(err)
-            .error('Could not delete entity')
-          res.status(404).send(err.message)
-        }
       })
     )
